requirements.txt
python-docx
openai
python-dotenv
flet
pyinstaller
